+++ src/main/java/cs698/giraph/kmode/InputFormat.java
@@ -53,9 +53,9 @@ public class InputFormat extends TextVertexInputFormat<LongWritable, NodeState, 
 			Text line = getRecordReader().getCurrentValue();
 			String[] parts = line.toString().split(",");
 			if(parts.length == 5) {
-				double [] data = new double[parts.length];
+				int [] data = new int[parts.length];
 				for(int i = 0; i< data.length; i++) {
-					data[i] = Double.parseDouble(parts[i]);
+					data[i] = Integer.parseInt(parts[i]);
 				}
 				LongWritable lwid = new LongWritable(id++);
 				NodeState value = new NodeState(new PointWritable(data));
